Editing tagline of toastman tomato article to be shorter
In Google search results, I have heard various maximum character counts
thrown around, but with Google Plus attribution and inclusion of the article
date, these can be off by a bit.  As such, I have shortened the tagline for
this article to try to get it all in, and not have anything replaced by
ellipsis.